+++ ChangeLog.rst
@@ -7,7 +7,7 @@
 * Feature #213: `images_as_html` config option to always generate an `img` html tag. preserves "height", "width" and "alt" if possible.
 * Remove support for end-of-life Pythons. Now requires Python 2.7 or 3.4+.
 
-2018.9.1
+2018.1.9
 ========
 ----